+++ src/ios/firebase/FirebaseCore.framework/Headers/FirebaseCore.h
@@ -1,0 +1,6 @@
+// Generated umbrella header for FirebaseCore.
+
+#import "FIRAnalyticsConfiguration.h"
+#import "FIRApp.h"
+#import "FIRConfiguration.h"
+#import "FIROptions.h"
